# public/apidoc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APD-API",
        "description": "An API to perform air pressure drop calculation through projects and user account",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://aeraulic.io/api/V1",
            "description": "Apd calculator API server"
        }
    ],
    "paths": {
...
        "/apd/users/{id}/projects": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get all project",
                "description": "Get all the projects from an user.",
                "operationId": "ce054671e6e57121217c4e804c58a217",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success to get all projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/project"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/apd/users/{userId}/projects/{ProjectId}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get one project",
                "description": "Get one project with his id from an user account.",
                "operationId": "0ec8d1d20f00b6647953142ebb9c8466",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/userId"
                    },
                    {
                        "$ref": "#/components/parameters/projectId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success to get one project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/project"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/apd/users/{userId}/projects": {
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Add a project",
                "description": "Add a project in the user account.",
                "operationId": "f058f85080d02d5f744d7994b33227c4",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/userId"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/addProject"
                },
                "responses": {
                    "200": {
                        "description": "Success to add a project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "Project's id",
                                                    "type": "integer",
                                                    "example": 21
                                                },
                                                "name": {
                                                    "title": "name",
                                                    "description": "Project's name",
                                                    "type": "string",
                                                    "example": "project A"
                                                },
                                                "userId": {
                                                    "title": "user id",
                                                    "description": "Project's associated user id",
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "generalAltitude": {
                                                    "title": "general altitude",
                                                    "description": "Project's altitude below sea level, all duct networks and duct sections are dependant of this property - in meter (m)",
                                                    "type": "integer",
                                                    "example": 800
                                                },
                                                "generalTemperature": {
                                                    "title": "general temperature",
                                                    "description": "Project's temperature, all duct networks and duct sections are dependant of this property - in degrees Celsius (°C)",
                                                    "type": "number",
                                                    "example": 18.2
                                                },
                                                "ductNetworks": {
                                                    "title": "duct networks",
                                                    "description": "All duct networks associated of this project",
                                                    "type": "array",
                                                    "items": {},
                                                    "example": "[]"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/unprocessableEntity"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/apd/users/{userId}/projects/{projectId}": {
            "delete": {
                "tags": [
                    "Project"
                ],
                "summary": "Remove a project",
                "description": "Remove a project in the user account. All duct networks and duct section associated will be deleted too.",
                "operationId": "49a5b7812d8c1b14b897a0678e27f678",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/userId"
                    },
                    {
                        "$ref": "#/components/parameters/projectId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success to remove a project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Project with name \"Project 1\" has been deleted, all associated duct networks and duct sections has been deletes too."
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/project"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Project"
                ],
                "summary": "Update a project",
                "description": "Update a project in the user account",
                "operationId": "34ffae8b97fe918afdb45e414b474579",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/userId"
                    },
                    {
                        "$ref": "#/components/parameters/projectId"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/updateProject"
                },
                "responses": {
                    "200": {
                        "description": "Success to update a project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/project"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/notFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/unprocessableEntity"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "project": {
                "title": "project",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Project's id",
                        "type": "integer",
                        "example": 21
                    },
                    "name": {
                        "title": "name",
                        "description": "Project's name",
                        "type": "string",
                        "example": "project A"
                    },
                    "userId": {
                        "title": "user id",
                        "description": "Project's associated user id",
                        "type": "integer",
                        "example": 10
                    },
                    "generalAltitude": {
                        "title": "general altitude",
                        "description": "Project's altitude below sea level, all duct networks and duct sections are dependant of this property - in meter (m)",
                        "type": "integer",
                        "minimum": 0,
                        "example": 800
                    },
                    "generalTemperature": {
                        "title": "general temperature",
                        "description": "Project's temperature, all duct networks and duct sections are dependant of this property - in degrees Celsius (°C)",
                        "type": "number",
                        "format": "float",
                        "example": 18.2
                    },
                    "ductNetworks": {
                        "title": "duct networks",
                        "description": "All duct networks associated of this project",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ductNetwork"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "notFound": {
                "description": "Ressource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Error"
                                },
                                "content": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": "id"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Ressource doesn't exist with this id."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "unprocessableEntity": {
                "description": "Unable to process the contained instructions",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Error"
                                },
                                "content": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": "shape"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Value \"rectangulard\" is not an element of the valid values: circular, rectangular."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Ressource's id",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "userId": {
                "name": "userId",
                "in": "path",
                "description": "User's id",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "projectId": {
                "name": "projectId",
                "in": "path",
                "description": "Project's id",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "requestBodies": {
...
            "addProject": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "title": "name",
                                    "description": "Name of the project",
                                    "type": "string",
                                    "example": "project B"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "updateProject": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "title": "name",
                                    "description": "Name of the project",
                                    "type": "string",
                                    "example": "project B",
                                    "nullable": true
                                },
                                "generalAltitude": {
                                    "title": "general altitude",
                                    "description": "Project's altitude below sea level, all duct networks and duct sections are dependant of this property - in meter (m",
                                    "type": "integer",
                                    "minimum": 0,
                                    "example": 800,
                                    "nullable": true
                                },
                                "generalTemperature": {
                                    "title": "general temperature",
                                    "description": "Project's temperature, all duct networks and duct sections are dependant of this property - in degrees Celsius (°C)",
                                    "type": "number",
                                    "format": "float",
                                    "example": 18.2,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}